Update in countBatteriesByUsage inside Main.java
I have made neccessary changes to get output of the code in the countBatteriesByUsage method. I have used a for loop to iterate through the cycles array by using the length. Then this for loop will check if our cycles is low, medium or high based on the cycles array. I have used the values given in the description to check the count of the  cycles. Once the entire array length is iterated the function will return the respective counts. The counts are accessed from the static class and finally updated in the loop. This is then verified in the testBucketingByNumberOfCycles and finally we can see our desired output.